Enable Windows CI //source/exe:envoy-static build (#10293)
Description:
 - Build and CI tooling to get windows compiling
 - (Reintroduces an experimental ci/windows_build_steps.ps1)
 - Source changes required for Windows compilation on msvc
 - Pick up a flavor of nameser.h for Windows only from c-ares dependency
 - Excludes all source fixes unrelated to compilation passing on Windows

Risk Level: Low
Testing: local on linux/gcc windows/msvc
Docs Changes: n/a
Release Notes: n/a

+++ source/extensions/filters/http/common/compressor/compressor.cc
@@ -208,7 +208,8 @@ CompressorFilter::chooseEncoding(const Http::ResponseHeaderMap& headers) const {
 
   // Find all encodings accepted by the user agent and adjust the list of allowed compressors.
   for (const auto token : StringUtil::splitToken(*accept_encoding_, ",", false /* keep_empty */)) {
-    EncPair pair = std::make_pair(StringUtil::trim(StringUtil::cropRight(token, ";")), 1);
+    EncPair pair =
+        std::make_pair(StringUtil::trim(StringUtil::cropRight(token, ";")), static_cast<float>(1));
     const auto params = StringUtil::cropLeft(token, ";");
     if (params != token) {
       const auto q_value = StringUtil::cropLeft(params, "=");
@@ -245,7 +246,7 @@ CompressorFilter::chooseEncoding(const Http::ResponseHeaderMap& headers) const {
 
   // Find intersection of encodings accepted by the user agent and provided
   // by the allowed compressors and choose the one with the highest q-value.
-  EncPair choice{Http::Headers::get().AcceptEncodingValues.Identity, 0};
+  EncPair choice{Http::Headers::get().AcceptEncodingValues.Identity, static_cast<float>(0)};
   for (const auto pair : pairs) {
     if ((pair.second > choice.second) &&
         (allowed_compressors.count(std::string(pair.first)) ||
